Finish frontend refactor
