+++ frontend/src/lib/AuthModal/AuthModal.svelte
