+++ frontend/src/lib/AuthModal/Error.svelte
