+++ frontend/src/lib/AuthModal/Logout.svelte
