+++ frontend/src/lib/AuthModal/helpers.js
