+++ frontend/src/lib/PhotoModal.svelte
