+++ frontend/src/routes/admin/inventory/components/ClaimItemModal.svelte
@@ -1,7 +1,7 @@
 <script>
 	import { onMount } from 'svelte';
-	import { formatDate } from '$lib/root/helpers.js';
-	import { claimItem } from './helpers.js';
+	import { formatDate } from '$lib/helpers.js';
+	import { claimItem } from '$lib/api_calls/admin_items.js';
 	import { refreshItems } from '$lib/stores/items';
 	import { users, refreshUsers } from '$lib/stores/users';
 	let { showClaimModal = $bindable(false), selectedItem = null } = $props();
