+++ frontend/src/routes/admin/inventory/components/ClaimedItemsAdmin.svelte
@@ -1,7 +1,7 @@
 <script>
-	import { formatDate } from '../../../lib/root/helpers.js';
+	import { formatDate } from '$lib/helpers.js';
 	import { claimedItems } from '$lib/stores/items';
-	import { removeClaimedItem } from '../../../lib/admin/inventory/helpers.js';
+	import { removeClaimedItem } from '$lib/api_calls/admin_items.js';
 	import { refreshItems } from '$lib/stores/items';
 
 	async function handleRemove(itemId) {
