+++ frontend/src/routes/admin/inventory/components/ItemsAdmin.svelte
@@ -1,9 +1,10 @@
 <script>
 	import { items } from '$lib/stores/items';
-	import { formatDate } from './helpers';
+	import { formatDate } from '$lib/helpers.js';
 	import EditItemModal from './EditItemModal.svelte';
 	import ClaimItemModal from './ClaimItemModal.svelte';
-	import PhotoModal from '../../../root_components/PhotoModal.svelte';
+	import PhotoModal from '$lib/PhotoModal.svelte';
+
 	let selectedItem = $state(null);
 	let showEditModal = $state(false);
 	let showClaimModal = $state(false);
